gamecenter/views/GameView.py
from rest_framework import viewsets
from gamecenter.models import Game
from gamecenter.serializers import GameSerializer
from rest_framework.filters import SearchFilter, OrderingFilter
from django_filters.rest_framework import DjangoFilterBackend

class GameViewSet(viewsets.ModelViewSet):
    queryset = Game.objects.all()
    serializer_class = GameSerializer

    filter_backends = [SearchFilter, DjangoFilterBackend]
    search_fields = ['name']
    filterset_fields = ['gender', 'game_material_type','is_active']